Removed all implementations except RemovableRBMIntKeyLookupStore, which is renamed as RBMIntKeyLookupStore
