+++ server/src/main/java/org/opensearch/indices/RBMSizeEstimator.java
@@ -34,8 +34,9 @@ package org.opensearch.indices;
 
 /**
  * A class used to estimate roaring bitmap memory sizes (and hash set sizes).
- * An instance is made with a particular modulo to avoid recomputing
- * values.
+ * Values based on experiments with adding randomly distributed integers, which matches the use case for KeyLookupStore.
+ * In this use case, true values are much higher than an RBM's self-reported size, especially for small RBMs: see
+ * https://github.com/RoaringBitmap/RoaringBitmap/issues/257
  */
 public class RBMSizeEstimator {
     public static final int BYTES_IN_MB = 1048576;
@@ -44,15 +45,18 @@ public class RBMSizeEstimator {
     public static final double bufferMultiplier = 1.5;
     public static final double intercept = 2.9;
 
-
     RBMSizeEstimator() {}
 
     public static long getSizeInBytes(int numEntries) {
+        // Based on a linear fit in log-log space, so that we minimize the error as a proportion rather than as
+        // an absolute value. Should be within ~50% of the true value at worst, and should overestimate rather
+        // than underestimate the memory usage
         return (long) ((long) Math.pow(numEntries, slope) * (long) Math.pow(10, intercept) * bufferMultiplier);
     }
 
     public static int getNumEntriesFromSizeInBytes(long sizeInBytes) {
-        // This function has some precision issues especially when composed with its inverse: numEntries = getNumEntriesFromSizeInBytes(getSizeInBytes(numEntries))
+        // This function has some precision issues especially when composed with its inverse:
+        // numEntries = getNumEntriesFromSizeInBytes(getSizeInBytes(numEntries))
         // In this case the result can be off by up to a couple percent
         // However, this shouldn't really matter as both functions are based on memory estimates with higher errors than a couple percent
         // and this composition won't happen outside of tests
